test/clib: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39295 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ test/clib/chown.c
@@ -16,15 +16,15 @@ int main()
     close(fd);
 
     TEST((stat(testfilename, &buf) != -1));
-    printf("owner %d group %d\n", buf.st_uid, buf.st_gid);
+    printf("owner %d group %d\n", (int)buf.st_uid, (int)buf.st_gid);
 
     TEST((chown(testfilename, 1000, 1001) != -1));
     TEST((stat(testfilename, &buf) != -1));
-    printf("owner %d group %d\n", buf.st_uid, buf.st_gid);
+    printf("owner %d group %d\n", (int)buf.st_uid, (int)buf.st_gid);
 
     TEST((chown(testfilename, 1001, 1000) != -1));
     TEST((stat(testfilename, &buf) != -1));
-    printf("owner %d group %d\n", buf.st_uid, buf.st_gid);
+    printf("owner %d group %d\n", (int)buf.st_uid, (int)buf.st_gid);
 
     cleanup();
     return OK;
